# gh-91851: Micro optimizations for Fraction's arithmetic