--- a/Misc/NEWS.d/next/Library/2022-04-23-08-12-14.gh-issue-91851.Jd47V6.rst
+++ b/Misc/NEWS.d/next/Library/2022-04-23-08-12-14.gh-issue-91851.Jd47V6.rst
@@ -1,0 +1,1 @@
+Optimize the :class:`~fractions.Fraction` arithmetics for small components.