This apply some suggestions from the #24779 review.

I've tested also direct imports (e.g. `from math import gcd`), [suggested](https://github.com/python/cpython/pull/24779#issuecomment-801449419) by @tim-one, but the difference doesn't seems to be noticeable.

Resolves #91851

<!-- gh-issue-number: gh-91851 -->
* Issue: gh-91851
<!-- /gh-issue-number -->

Benchmarks (taken from the second commit msg):

That makes arithmetics for small components just as fast
as before https://github.com/python/cpython/pull/24779, except for a mixed case (e.g.
Fraction + int):

On master:
```
  $ ./python -m timeit -s 'from fractions import Fraction as F' \
    -s 'a = F(7, 8)' -s 'b = F(3, 4)' 'a + b'
  20000 loops, best of 5: 12.4 usec per loop
  $ ./python -m timeit -s 'from fractions import Fraction as F' \
    -s 'a = F(7, 8)' -s 'b = 3' 'a + b'
  20000 loops, best of 5: 10.3 usec per loop
  $ ./python -m timeit -s 'from fractions import Fraction as F' \
    -s 'a = 3' -s 'b = F(3, 4)' 'a + b'
  20000 loops, best of 5: 12.6 usec per loop
```

With the patch:
```
  $ ./python -m timeit -s 'from fractions import Fraction as F' \
    -s 'a = F(7, 8)' -s 'b = F(3, 4)' 'a + b'
  20000 loops, best of 5: 9.98 usec per loop
  $ ./python -m timeit -s 'from fractions import Fraction as F' \
    -s 'a = F(7, 8)' -s 'b = 3' 'a + b'
  20000 loops, best of 5: 15.6 usec per loop
  $ ./python -m timeit -s 'from fractions import Fraction as F' \
    -s 'a = 3' -s 'b = F(3, 4)' 'a + b'
  20000 loops, best of 5: 16.6 usec per loop
```