@skirpichev 2023-01-04
@hauntsaninja, actually there were benchmarks (second commit message).  Now copied to the pr's body, thanks.