@eendebakpt 2023-01-04
> @hauntsaninja, actually there were benchmarks (second commit message). Now copied to the pr's body, thanks.

@skirpichev @hauntsaninja The PR indeed makes addition of Fraction and int slower. The regression is in this https://github.com/python/cpython/pull/25518/files#diff-f561eb7eb97f832b2698837f52c2c2cf23bdb0b31c69cf1f6aaa560280993316R398 change, which forces the `int` types to convert from `int` to `Fraction`. In the current main that converstion is not required since the `int` type has properties `.denominator` and `.numerator` just like the Fraction type. Reverting that change would force all all `monomorphic_operator` functions to make only use of `.numerator` and `.denominator` (and not of `._denominator`). I do not see how to make the construction of `Fraction` from `int` much faster, there is already a fast pass in the `Fraction` constructor.

The optimizations in `__round__`, `__ceil__` and `__floor__` are still good.