@eendebakpt 2023-01-04
Also here https://github.com/python/cpython/pull/25518/files#diff-f561eb7eb97f832b2698837f52c2c2cf23bdb0b31c69cf1f6aaa560280993316R414 there is a regression for the case `1+Fraction(2)` and other cases `r+Fraction(2)` where `r` is of `Rational` type.