@skirpichev 2023-01-05
@eendebakpt, arithmetic optimization assume that the hot path is `Fraction op Fraction`.  I don't see how to avoid speed regression here.  Perhaps, only with the cost of using public properties for the Rational subtypes and private - only for the Fraction type.  Any thoughts?

> The optimizations in __round__, __ceil__ and __floor__ are still good.
Should I create a separate pr to get this merged?