@hauntsaninja 2023-01-05
If you create a separate PR for round, ceil and floor, I can merge it since that's uncontroversial :-)