@hauntsaninja 2023-01-07
Thanks for splitting off the other PR and rebasing this one. I don't have good enough insight into fractions to know if this tradeoff is worth it, e.g. if Fraction + int is as common as Fraction + Fraction, this is not a win.