@eendebakpt 2023-01-07
The change to use properties for `numerator` and `deniminator` was done 15 years ago. From the comments I gather the reason was to make them read-only to prevent users from changing them. Reverting that change to make this PR faster would break backwards compatibility.