@mdickinson 2023-01-07
The changes LGTM - I find the new code to be cleaner (as well as fast for the `Fraction op Fraction` case).