@skirpichev 2023-01-08
@eendebakpt 
> The change to use properties for numerator and deniminator was done 15 years ago.

It seems, 7736b5becd2 is that commit, correct?  I don't see mentioned comments in the related discussion (#46023).

> break backwards compatibility

Could you give an example?