@eendebakpt 2023-01-08
> @eendebakpt
> 
> > The change to use properties for numerator and deniminator was done 15 years ago.
> 
> It seems, [7736b5b](https://github.com/python/cpython/commit/7736b5becd273cb271c55fcce9155677a381c4f6) is that commit, correct? I don't see mentioned comments in the related discussion (#46023).
> 
> > break backwards compatibility
> 
> Could you give an example?

The commit is https://github.com/python/cpython/commit/400adb030a78c3fcb1eb458b25c990449fbf3c93

Such a change would impact subclasses of Fraction that directly set '._numerator' . I am not aware of such subclasses, but they can exist.