@hauntsaninja 2023-01-08
Thanks for the PR! And yes, definitely nice that monomorphic_operator is now more monomorphic