@skirpichev 2023-01-08
Thanks to all for reviews.

> Such a change would impact subclasses of Fraction that directly set '._numerator' .

@eendebakpt  You can do this as long as you respect constraints on these private attributes.  I.e. they are essentially numerator/denominator properties of the numbers.Rational ("The [numerator](https://docs.python.org/3/library/numbers.html#numbers.Rational.numerator) and [denominator](https://docs.python.org/3/library/numbers.html#numbers.Rational.denominator) values should be instances of [Integral](https://docs.python.org/3/library/numbers.html#numbers.Integral) and should be in lowest terms with [denominator](https://docs.python.org/3/library/numbers.html#numbers.Rational.denominator) positive.").  If you ignore this - some methods of the Fraction class will be broken anyway, starting from the `__eq__()`.

I hardly can imagine benefits of subclassing the Fraction.  E.g. if you want to keep components in an unnormalized form - most arithmetic methods should be revised.